handle_{saving,loading} to Session
