+++ src/pytest_benchmark/session.py
@@ -3,7 +3,6 @@ from __future__ import print_function
 
 import pytest
 
-from pytest_benchmark.reporting import Reporter
 from .fixture import statistics
 from .fixture import statistics_error
 from .logger import Logger
@@ -16,6 +15,7 @@ from .utils import get_machine_id
 from .utils import load_storage
 from .utils import load_timer
 from .utils import short_filename
+from .utils import safe_dumps
 
 
 class PerformanceRegression(pytest.UsageError):
@@ -33,13 +33,10 @@ class BenchmarkSession(object):
         self.performance_regressions = []
         self.benchmarks = []
         self.machine_id = get_machine_id()
-        self.reporter = Reporter(
-            config=config,
-            storage=load_storage(
-                config.getoption("benchmark_storage"),
-                logger=self.logger,
-                default_machine_id=self.machine_id
-            )
+        self.storage = load_storage(
+            config.getoption("benchmark_storage"),
+            logger=self.logger,
+            default_machine_id=self.machine_id
         )
         self.options = dict(
             min_time=SecondsDecimal(config.getoption("benchmark_min_time")),
@@ -85,6 +82,11 @@ class BenchmarkSession(object):
                 "Can't have both --benchmark-only and --benchmark-disable options. Note that --benchmark-disable is "
                 "automatically activated if xdist is on or you're missing the statistics dependency.")
         self.group_by = config.getoption("benchmark_group_by")
+        self.save = config.getoption("benchmark_save")
+        self.autosave = config.getoption("benchmark_autosave")
+        self.save_data = config.getoption("benchmark_save_data")
+        self.json = config.getoption("benchmark_json")
+        self.compare = config.getoption("benchmark_compare")
         self.compare_fail = config.getoption("benchmark_compare_fail")
         self.name_format = NAME_FORMATTERS[config.getoption("benchmark_name")]
 
@@ -122,9 +124,81 @@ class BenchmarkSession(object):
                 flat_bench["source"] = compared and "NOW"
                 yield flat_bench
 
+    def save_json(self, output_json):
+        with self.json as fh:
+            fh.write(safe_dumps(output_json, ensure_ascii=True, indent=4).encode())
+        self.logger.info("Wrote benchmark data in: %s" % self.json, purple=True)
+
+    def handle_saving(self):
+        save = self.benchmarks and self.save or self.autosave
+        if save or self.json:
+            commit_info = self.config.hook.pytest_benchmark_generate_commit_info(config=self.config)
+            self.config.hook.pytest_benchmark_update_commit_info(config=self.config, commit_info=commit_info)
+
+        if self.json:
+            output_json = self.config.hook.pytest_benchmark_generate_json(
+                config=self.config,
+                benchmarks=self.benchmarks,
+                include_data=True,
+                machine_info=self.machine_info,
+                commit_info=commit_info,
+            )
+            self.config.hook.pytest_benchmark_update_json(
+                config=self.config,
+                benchmarks=self.benchmarks,
+                output_json=output_json,
+            )
+            self.save_json(output_json)
+
+        if save:
+            output_json = self.config.hook.pytest_benchmark_generate_json(
+                config=self.config,
+                benchmarks=self.benchmarks,
+                include_data=self.save_data,
+                machine_info=self.machine_info,
+                commit_info=commit_info,
+            )
+            self.config.hook.pytest_benchmark_update_json(
+                config=self.config,
+                benchmarks=self.benchmarks,
+                output_json=output_json,
+            )
+            self.storage.save(output_json, save)
+
+    def handle_loading(self):
+        compared_mapping = {}
+        if self.compare:
+            if self.compare is True:
+                compared_benchmarks = list(self.storage.load())[-1:]
+            else:
+                compared_benchmarks = list(self.storage.load(self.compare))
+
+            if not compared_benchmarks:
+                msg = "Can't compare. No benchmark files in %r" % str(self.storage)
+                if self.compare is True:
+                    msg += ". Can't load the previous benchmark."
+                    code = "BENCHMARK-C2"
+                else:
+                    msg += " match %r." % self.compare
+                    code = "BENCHMARK-C1"
+                self.logger.warn(code, msg, fslocation=self.storage.location)
+
+            for path, compared_benchmark in compared_benchmarks:
+                self.config.hook.pytest_benchmark_compare_machine_info(
+                    config=self.config,
+                    benchmarksession=self,
+                    machine_info=self.machine_info,
+                    compared_benchmark=compared_benchmark,
+                )
+                compared_mapping[path] = dict(
+                    (bench['fullname'], bench) for bench in compared_benchmark['benchmarks']
+                )
+                self.logger.info("Comparing against benchmarks from: %s" % path)
+        self.compared_mapping = compared_mapping
+
     def finish(self):
-        self.reporter.handle_saving(self.benchmarks, self.machine_info)
-        self.compared_mapping = self.reporter.handle_loading(self.machine_info)
+        self.handle_saving()
+        self.handle_loading()
         prepared_benchmarks = list(self.prepare_benchmarks())
         if prepared_benchmarks:
             self.groups = self.config.hook.pytest_benchmark_group_stats(
@@ -173,6 +247,7 @@ class BenchmarkSession(object):
                         tr.write("\n")
                     tr.write_line("ncalls\ttottime\tpercall\tcumtime\tpercall\tfilename:lineno(function)")
                     for function_info in benchmark["cprofile"]:
-                        line = "{ncalls_recursion}\t{tottime:.{prec}f}\t{tottime_per:.{prec}f}\t{cumtime:.{prec}f}\t{cumtime_per:.{prec}f}\t{function_name}".format(prec=4, **function_info)
+                        line = "{ncalls_recursion}\t{tottime:.{prec}f}\t{tottime_per:.{prec}f}\t{cumtime:.{prec}f}\t{cumtime_per:.{prec}f}\t{function_name}".format(
+                            prec=4, **function_info)
                         tr.write_line(line)
                     tr.write("\n")
