+++ tests/test_elasticsearch_storage.py
@@ -15,7 +15,6 @@ from pytest_benchmark.plugin import BenchmarkSession
 from pytest_benchmark.plugin import pytest_benchmark_compare_machine_info
 from pytest_benchmark.plugin import pytest_benchmark_generate_json
 from pytest_benchmark.plugin import pytest_benchmark_group_stats
-from pytest_benchmark.reporting import Reporter
 from pytest_benchmark.storage.elasticsearch import ElasticsearchStorage
 
 try:
@@ -58,31 +57,19 @@ class MockStorage(ElasticsearchStorage):
         self._es_hosts = self._es_index = self._es_doctype = 'mocked'
         self.logger = logger
 
-class MockReporter(Reporter):
-    def __init__(self, config):
-        self.verbose = False
-        self.logger = logger
-        self.config = config
-        self.performance_regressions = []
-        self.benchmarks = []
-        self.machine_id = "FoobarOS"
-        self.storage = MockStorage()
-        self.compare = '0001'
-        self.save = self.autosave = self.json = False
-        self.elasticsearch_hosts = ["localhost:9200"]
-        self.elasticsearch_index = "benchmark"
-        self.elasticsearch_doctype = "benchmark"
-
 
 class MockSession(BenchmarkSession):
     def __init__(self):
+        self.verbose = False
         self.histogram = True
         self.benchmarks = []
         self.performance_regressions = []
         self.sort = u"min"
+        self.compare = '0001'
         self.logger = logging.getLogger(__name__)
         self.machine_id = "FoobarOS"
         self.machine_info = {'foo': 'bar'}
+        self.save = self.autosave = self.json = False
         self.options = {
             'min_rounds': 123,
             'min_time': 234,
@@ -102,7 +89,7 @@ class MockSession(BenchmarkSession):
         self.elasticsearch_host = "localhost:9200"
         self.elasticsearch_index = "benchmark"
         self.elasticsearch_doctype = "benchmark"
-        self.report_backend = MockReporter(self.config)
+        self.storage = MockStorage()
         self.group_by = 'group'
         self.columns = ['min', 'max', 'mean', 'stddev', 'median', 'iqr',
                         'outliers', 'rounds', 'iterations']
@@ -152,7 +139,7 @@ def make_logger(sess):
         info=lambda text, **opts: output.write(force_text(text) + u'\n'),
         error=lambda text: output.write(force_text(text) + u'\n'),
     )
-    sess.report_backend.logger = Namespace(
+    sess.storage.logger = Namespace(
         warn=lambda code, text, **opts: output.write(u"%s: %s %s\n" % (code, force_text(text), opts)),
         info=lambda text, **opts: output.write(force_text(text) + u'\n'),
         error=lambda text: output.write(force_text(text) + u'\n'),
@@ -173,12 +160,12 @@ def logger_output(sess):
 @freeze_time("2015-08-15T00:04:18.687119")
 def test_handle_saving(sess, logger_output, monkeypatch):
     monkeypatch.setattr(plugin, '__version__', '2.5.0')
-    sess.report_backend.save = "commitId"
-    sess.report_backend.autosave = True
-    sess.report_backend.json = None
-    sess.report_backend.save_data = False
-    sess.report_backend.handle_saving(sess.benchmarks, sess.machine_info)
-    sess.report_backend.storage._es.index.assert_called_with(
+    sess.save = "commitId"
+    sess.autosave = True
+    sess.json = None
+    sess.save_data = False
+    sess.handle_saving()
+    sess.storage._es.index.assert_called_with(
         index='mocked',
         doc_type='mocked',
         body=ES_DATA,
